test: move tests-specific methods to actual tests
